User score increase development completed
User score increase development completed

+++ src/main/java/com/whisper/dto/UserDTO.java
@@ -10,6 +10,6 @@ import java.util.Set;
 @Data
 public class UserDTO {
     private String username;
-    private Integer userPoint;
+    private Double userPoint;
     private String authorities;
 }
